# Optimize `DateToString` converter