--- a/app/server/conversion/factory.go
+++ b/app/server/conversion/factory.go
@@ -1,0 +1,13 @@
+package conversion
+
+import (
+	"github.com/ydb-platform/fq-connector-go/app/config"
+)
+
+func NewCollection(cfg *config.TConversionConfig) Collection {
+	if cfg.UseUnsafeConverters {
+		return collectionUnsafe{}
+	}
+
+	return collectionDefault{}
+}