--- a/example.conf
+++ b/example.conf
@@ -21,3 +21,7 @@ paging {
     bytes_per_page: 4194304
     prefetch_queue_capacity: 2
 }
+
+conversion {
+    use_unsafe_converters: false
+}