Provided `unsafe`-based implementation of `DateToString` converter:

```
go test -bench="DateToString" -memprofile profile.out -benchmem -v ./app/server/conversion
(...)
cpu: 11th Gen Intel(R) Core(TM) i7-11850H @ 2.50GHz
BenchmarkDateToStringConverter
BenchmarkDateToStringConverter/Default
BenchmarkDateToStringConverter/Default-16               17300325                67.20 ns/op           16 B/op          1 allocs/op
BenchmarkDateToStringConverter/Unsafe
BenchmarkDateToStringConverter/Unsafe-16                25947836                43.99 ns/op           16 B/op          1 allocs/op
```


Benchmarking request: `SELECT l_shipdate FROM lineitem` in PostgreSQL 15.3. Throughput measurements:

Before
* 60,4 MB/sec
* 3.95 * 10^6 Rows/sec

After:
* 69.8 MB/sec **(+15.6%)**
* 4.57 * 10^6 Rows/sec **(+15.7%)**